Add an option to use the CDN instead of the REST API
In some servers, the REST API cannot be used because of the restrictions of the fsockopen by the provider. In that case it is now possible to use the CDN javascript provided by http://www.websequencediagrams.com/service.js instead.
There is now a new configuration option in the administration of the plugin.